+++ src/effects/visible/index.js
@@ -40,7 +40,7 @@ class Visible {
    *   @prop {number} max
    *  }
    * }
-   * @return {void|boolean} - false if init errors
+   * @return {void|boolean} - False if init errors
    */
 
   constructor (args) {